winter_utils

Trait AsBytes

pub trait AsBytes {
    // Required method
    fn as_bytes(&self) -> &[u8] ;
}

Defines a zero-copy representation of Self as a sequence of bytes.

Required Methods

fn as_bytes(&self) -> &[u8]

Returns a byte representation of self.

This method is intended to re-interpret the underlying memory as a sequence of bytes, and thus, should be zero-copy.