winter_utils

Trait Randomizable

pub trait Randomizable: Sized {
    const VALUE_SIZE: usize;

    // Required method
    fn from_random_bytes(source: &[u8]) -> Option<Self>;
}

Defines how Self can be read from a sequence of random bytes.

Required Associated Constants

const VALUE_SIZE: usize

Size of Self in bytes.

This is used to determine how many bytes should be passed to the from_random_bytes() function.

Required Methods

fn from_random_bytes(source: &[u8]) -> Option<Self>

Returns Self if the set of bytes forms a valid value, otherwise returns None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.